added authorization checks for delete and update endpoints

+++ controllers/postController.js
@@ -1,7 +1,7 @@
 import Post from "../models/post.js";
 export const getPosts = async (req, res) => {
     try{
-        const posts = await Post.find();
+        const posts = await Post.find().populate("author", "username");
         if (!posts) return res.status(404).json({error: "No Posts Found!"});
         res.json(posts);
     } catch (e){
@@ -12,7 +12,7 @@ export const getPosts = async (req, res) => {
 export const getPost = async (req, res) => {
     try{
         const id = req.params.id;
-        const post = await Post.findById(id);
+        const post = await Post.findById(id).populate("author", "username");
         if (!post) return res.status(404).json({error: "Post not Found!"});
         res.json(post);
     } catch (e) {
@@ -23,6 +23,7 @@ export const getPost = async (req, res) => {
 export const createPost = async (req, res) => {
     try{
         const post = req.body;
+        post.author = req.user.id;  // sets the userid to the author
         const newPost = await Post.create(post);
         if (newPost) res.status(201).json({message: "Post Created Successfully!"})
     } catch (e) {
@@ -32,10 +33,14 @@ export const createPost = async (req, res) => {
 
 export const deletePost = async(req, res) => {
     try{
-        const id = req.params.id;
-        const post = await Post.findByIdAndDelete(id);
-        if (!post) return res.status(404).json({error: "Couldn't find post"});
-        return res.status(204).send(post);
+      const id = req.params.id;
+      const post = await Post.findById(id);
+      if (!post) return res.status(404).json({ error: "Couldn't find post" });
+      if (post.author.toString() !== req.user.id) // checks if the post is owned by the user 
+         return res.status(403).send("can't delete don't have permission");
+
+      const deletedPost=await Post.findByIdAndDelete(id)
+      return res.status(204).send(deletedPost);
     } catch (e) {
         return res.status(500).json({error: e.message})
     }
@@ -47,6 +52,8 @@ export const updatePost = async (req, res) =>{
         const id = req.params.id;
         const post = await Post.findById(id);
         if (!post) return res.status(404).json({error: "Couldn't find post"});
+        if (post.author.toString() !== req.user.id) // checks if the post is owned by the user
+              return res.status(403).send("can't update don't have permission");
         Object.keys(updatedPost).forEach( key => {
             post[key] = updatedPost[key];
         });
